+++ src/shared/components/checkbox/BgrCheckbox.tsx
